Fixes some bugs
