+++ js/main.js
@@ -1,13 +1,13 @@
-import {setSubmitForm} from './form.js';
 import {getData} from './api.js';
-import {photoDescriptions} from './thumbnails/render-thumbnails.js';
+import {renderPhotoDescriptions} from './thumbnails/render-thumbnails.js';
+import {configSoritng} from './thumbnails/sorting-thumbnails.js';
 import {setupPopup} from './popup-picture.js';
 import {appendDataError} from './util.js';
-import {configSoritng} from './thumbnails/sorting-thumbnails.js';
-//const imgFilters = document.querySelector('.img-filters');
+import {setSubmitForm} from './form/submit-form.js';
+
 getData()
   .then((photos) => {
-    photoDescriptions(photos);
+    renderPhotoDescriptions (photos);
     setupPopup (photos);
     configSoritng (photos);
   })
