+++ js/photo-effects/apply-effects.js
@@ -1,56 +1,12 @@
-import {SCALE, FILTERS_EFFECTS} from './data-for-effects';
+import {FILTERS_EFFECTS} from './data-for-effects';
 
 const form = document.querySelector('.img-upload__form');
-const scaleControlSmaller = form.querySelector('.scale__control--smaller');
-const scaleControlBigger = form.querySelector('.scale__control--bigger');
-const scaleControlValue = form.querySelector('.scale__control--value');
 const imgUploadPreview = form.querySelector('.img-upload__preview img');
 const imgUploadEffectLevel = form.querySelector('.img-upload__effect-level');
 const effectLevelValue = form.querySelector('.effect-level__value');
 const effectLevelSlider = form.querySelector('.effect-level__slider');
 const effectsList = form.querySelector('.effects__list');
 const effectsRadio = form.querySelectorAll('.effects__radio');
-const effectNone = form.querySelector('#effect-none');
-
-let newScaleControlValue = SCALE.DEFAULT;
-
-
-//Scale changing
-
-function decreaseSize () {
-  if (newScaleControlValue > SCALE.MIN) {
-    newScaleControlValue -= SCALE.STEP;
-    imgUploadPreview.style.transform = `scale(${newScaleControlValue / 100})`;
-    scaleControlValue.value = `${newScaleControlValue} %`;
-  }
-}
-
-scaleControlSmaller.addEventListener('click', (evt) => {
-  evt.preventDefault();
-  decreaseSize();
-});
-
-function increaseSize () {
-  if (newScaleControlValue < SCALE.MAX) {
-    newScaleControlValue += SCALE.STEP;
-    imgUploadPreview.style.transform = `scale(${newScaleControlValue / 100})`;
-    scaleControlValue.value = `${newScaleControlValue} %`;
-  }
-}
-const clearScale = function () {
-  scaleControlSmaller.disabled = false;
-  scaleControlBigger.disabled = false;
-  scaleControlValue.value = `${SCALE.DEFAULT} %`;
-  imgUploadPreview.style.transform = `scale(${SCALE.DEFAULT / 100})`;
-  effectNone.checked = true;
-};
-
-scaleControlBigger.addEventListener('click', (evt) => {
-  evt.preventDefault();
-  increaseSize();
-});
-
-//effects
 
 noUiSlider.create(effectLevelSlider, {
   range: {
@@ -79,12 +35,12 @@ effectLevelSlider.noUiSlider.on('update', () => {
 
 imgUploadEffectLevel.classList.add('hidden');
 
-const clearFilter = function () {
+const clearFilter = () => {
   imgUploadEffectLevel.classList.add('hidden');
   imgUploadPreview.style.filter = 'none';
 };
 
-const applyEffect = function () {
+const applyEffect = () => {
   effectsRadio.forEach((effectRadio) => {
     const effectValue = effectRadio.value;
     const effectName = FILTERS_EFFECTS[effectValue];
@@ -112,4 +68,4 @@ const applyEffect = function () {
 
 effectsList.addEventListener('change', applyEffect);
 
-export {clearFilter, clearScale};
+export {clearFilter};
