+++ js/thumbnails/sorting-thumbnails.js
@@ -1,7 +1,8 @@
+import {renderPhotoDescriptions} from './render-thumbnails';
 import {debounce} from '../util';
-import {photoDescriptions} from './render-thumbnails';
 
-const imgFilters = document.querySelector('.img-filters');
+const imgFiltersBlock = document.querySelector('.img-filters');
+
 const ACTIVE_BUTTON_CLASS = 'img-filters__button--active';
 
 let photos = [];
@@ -17,18 +18,19 @@ const SortingFuncs = {
   DISCUSSED: (a, b) => b.comments.length - a.comments.length
 };
 
-const RANDOM_PHOTOS_COUNT = 10;
+const RANDOM_PHOTOS_MAX_COUNT = 10;
 
 const clearThumbnails = () => {
   const thumbnails = document.querySelectorAll('a.picture');
   thumbnails.forEach((photoElement) => photoElement.remove());
 };
+
 const debounceRender = debounce((data) => {
   clearThumbnails();
-  photoDescriptions(data);
+  renderPhotoDescriptions(data);
 }, 500);
 
-function onFilterChange (evt) {
+const onFilterChange = (evt) => {
   const targetButton = evt.target;
   const activeButton = document.querySelector(`.${ACTIVE_BUTTON_CLASS}`);
   if (!targetButton.matches('button')) {
@@ -41,27 +43,28 @@ function onFilterChange (evt) {
   targetButton.classList.toggle(ACTIVE_BUTTON_CLASS);
   currentFilter = targetButton.getAttribute('id');
   sortingThumbnails();
-}
+};
 
 function sortingThumbnails () {
   let sortedPhotos = [];
+  switch (currentFilter) {
+    case Filters.DEFAULT:
+      sortedPhotos = photos;
+      break;
+    case Filters.RANDOM:
+      sortedPhotos = photos.toSorted(SortingFuncs.RANDOM).slice(0, RANDOM_PHOTOS_MAX_COUNT);
+      break;
+    case Filters.DISCUSSED:
+      sortedPhotos = photos.toSorted(SortingFuncs.DISCUSSED);
+  }
 
-  if (currentFilter === Filters.DEFAULT) {
-    sortedPhotos = photos;
-  }
-  if (currentFilter === Filters.RANDOM) {
-    sortedPhotos = photos.toSorted(SortingFuncs.RANDOM).slice(1, RANDOM_PHOTOS_COUNT);
-  }
-  if (currentFilter === Filters.DISCUSSED) {
-    sortedPhotos = photos.toSorted(SortingFuncs.DISCUSSED);
-  }
   debounceRender(sortedPhotos);
 }
 
-function configSoritng (data) {
-  imgFilters.classList.remove('img-filters--inactive');
-  imgFilters.addEventListener('click', (onFilterChange));
+const configSoritng = (data) => {
+  imgFiltersBlock.classList.remove('img-filters--inactive');
+  imgFiltersBlock.addEventListener('click', (onFilterChange));
   photos = data;
-}
+};
 
 export {configSoritng};
